requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
click==8.1.3
confusable-homoglyphs==3.2.0
Django==4.1.3
django-bootstrap4==22.2
django-crispy-forms==1.14.0
django-registration==2.4.1
django-urls==1.1.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
psycopg2==2.8.6
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.10.0
